Jest RTL conversions
